Add some more test + discover that valgrind works slightly better

+++ rop.sh
@@ -1,1 +1,6 @@
-./rop-fk $1 > a.tmp; cat a.tmp; rm a.tmp
+if ! command -v valgrind &> /dev/null
+then
+	./rop-fk $1 > a.tmp; cat a.tmp; rm a.tmp
+	exit
+fi
+valgrind ./rop-fk $1 2> /dev/null
